Split out fields in ViewPage and refactor code
