+++ src/components/Items/MyItemListing.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 // @material-ui/core components
 import { makeStyles } from '@material-ui/core/styles';
 import styles from 'assets/jss/Items/views/MyItemListing';
+import { Link } from 'react-router-dom';
 import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
@@ -19,20 +20,16 @@ import GridItem from 'components/MaterialKitComponents/Grid/GridItem';
 
 const useStyles = makeStyles(styles);
 
-const GetCollectionMsg = (type, time) => {
-  if (time === null) {
-    return null;
-  }
-
+const GetCollectionMsg = (type) => {
   let message = null;
   if (type === 'provide') {
-    message = `Being Collected: ${time}`;
+    message = 'Being Collected:';
   } else {
-    message = `To Collect: ${time}`;
+    message = 'To Collect:';
   }
 
   return (
-    <Typography variant='body2' color='textPrimary' gutterBottom>
+    <Typography variant='body2' color='textSecondary' gutterBottom>
       {message}
     </Typography>
   );
@@ -42,69 +39,211 @@ export default function MyItemListing({ type, myitem }) {
   const classes = useStyles();
 
   return (
-    <GridContainer spacing={2}>
+    <GridContainer spacing={1}>
       <GridItem>
-        <Paper className={classes.paper} spacing={2}>
-          <GridContainer spacing={2}>
+        <Paper className={classes.paper} spacing={1}>
+          <GridContainer spacing={1}>
             <Grid item>
-              <ButtonBase className={classes.image}>
-                <img
-                  className={classes.img}
-                  alt='complex'
-                  src={require('assets/img/purple-banana.jpg')}
-                />
-              </ButtonBase>
+              <Link to='/itemview' className={classes.link}>
+                <ButtonBase className={classes.image}>
+                  <img
+                    className={classes.img}
+                    alt='complex'
+                    src={require('assets/img/purple-banana.jpg')}
+                  />
+                </ButtonBase>
+              </Link>
             </Grid>
-            <GridContainer align='left' item xs={12} sm container>
-              <GridItem xs container direction='column' spacing={2}>
-                <GridItem xs={12}>
-                  <a href='#'>
-                    <Typography gutterBottom variant='body1'>
-                      <strong>{myitem.description}</strong>
-                    </Typography>
-                  </a>
-                  <GridContainer align='left' item xs={12} container>
-                    <GridItem xs={12} sm={6} md={6} lg={6}>
-                      <Typography
-                        variant='body2'
-                        color='textPrimary'
-                        gutterBottom
+            <GridContainer xs={12} sm item spacing={0} direction='column'>
+              <GridItem align='left' xs={12} className={classes.cell}>
+                <Link to='/itemview' className={classes.link}>
+                  <Typography gutterBottom variant='body1'>
+                    <strong>{myitem.description}</strong>
+                  </Typography>
+                </Link>
+                <GridContainer className={classes.container}>
+                  <GridItem
+                    xs={12}
+                    sm={6}
+                    md={6}
+                    lg={6}
+                    className={classes.cell}
+                  >
+                    <GridContainer
+                      direction='row'
+                      className={classes.container}
+                    >
+                      <GridItem
+                        xs={5}
+                        sm={5}
+                        sm={5}
+                        lg={5}
+                        className={classes.cell}
                       >
-                        Category:
-                        {myitem.category}
-                      </Typography>
-                    </GridItem>
-                    <GridItem xs={12} sm={6} md={6} lg={6}>
-                      <Typography
-                        variant='body2'
-                        color='textPrimary'
-                        gutterBottom
+                        <Typography
+                          variant='body2'
+                          color='textSecondary'
+                          gutterBottom
+                        >
+                          Category:
+                        </Typography>
+                      </GridItem>
+                      <GridItem
+                        xs={7}
+                        sm={7}
+                        sm={7}
+                        lg={7}
+                        align='left'
+                        className={classes.cell}
                       >
-                        Location:
-                        {myitem.location}
-                      </Typography>
-                    </GridItem>
-                  </GridContainer>
-                  <GridContainer align='left' item xs={12} container>
-                    <GridItem xs={12} sm={6} md={6} lg={6}>
-                      <Typography
-                        variant='body2'
-                        color='textPrimary'
-                        gutterBottom
+                        <Typography
+                          variant='body2'
+                          color='textPrimary'
+                          gutterBottom
+                          align='left'
+                        >
+                          {myitem.category}
+                        </Typography>
+                      </GridItem>
+                    </GridContainer>
+                  </GridItem>
+                  <GridItem
+                    xs={12}
+                    sm={6}
+                    md={6}
+                    lg={6}
+                    className={classes.cell}
+                  >
+                    <GridContainer
+                      direction='row'
+                      className={classes.container}
+                    >
+                      <GridItem
+                        xs={5}
+                        sm={5}
+                        sm={5}
+                        lg={5}
+                        className={classes.cell}
                       >
-                        <strong>
+                        <Typography
+                          variant='body2'
+                          color='textSecondary'
+                          gutterBottom
+                        >
+                          Location:
+                        </Typography>
+                      </GridItem>
+                      <GridItem
+                        xs={7}
+                        sm={7}
+                        sm={7}
+                        lg={7}
+                        align='left'
+                        className={classes.cell}
+                      >
+                        <Typography
+                          variant='body2'
+                          color='textPrimary'
+                          gutterBottom
+                          align='left'
+                        >
+                          {myitem.location}
+                        </Typography>
+                      </GridItem>
+                    </GridContainer>
+                  </GridItem>
+                </GridContainer>
+                <GridContainer className={classes.container}>
+                  <GridItem
+                    xs={12}
+                    sm={6}
+                    md={6}
+                    lg={6}
+                    className={classes.cell}
+                  >
+                    <GridContainer
+                      direction='row'
+                      className={classes.container}
+                    >
+                      <GridItem
+                        xs={5}
+                        sm={5}
+                        sm={5}
+                        lg={5}
+                        className={classes.cell}
+                      >
+                        <Typography
+                          variant='body2'
+                          color='textSecondary'
+                          gutterBottom
+                        >
                           Expires:
+                        </Typography>
+                      </GridItem>
+                      <GridItem
+                        xs={7}
+                        sm={7}
+                        sm={7}
+                        lg={7}
+                        align='left'
+                        className={classes.cell}
+                      >
+                        <Typography
+                          variant='body2'
+                          color='textPrimary'
+                          gutterBottom
+                          align='left'
+                        >
                           {myitem.expiryDate}
-                        </strong>
-                      </Typography>
-                    </GridItem>
-                    <GridItem xs={12} sm={6} md={6} lg={6}>
-                      {GetCollectionMsg(type, myitem.preferredCollectStartTime)}
-                    </GridItem>
-                  </GridContainer>
-                </GridItem>
+                        </Typography>
+                      </GridItem>
+                    </GridContainer>
+                  </GridItem>
+                  <GridItem
+                    xs={12}
+                    sm={6}
+                    md={6}
+                    lg={6}
+                    className={classes.cell}
+                  >
+                    <GridContainer
+                      direction='row'
+                      className={classes.container}
+                    >
+                      <GridItem
+                        xs={5}
+                        sm={5}
+                        sm={5}
+                        lg={5}
+                        className={classes.cell}
+                      >
+                        {myitem.preferredCollectStartTime !== null &&
+                          GetCollectionMsg(type)}
+                      </GridItem>
+                      <GridItem
+                        xs={7}
+                        sm={7}
+                        sm={7}
+                        lg={7}
+                        align='left'
+                        className={classes.cell}
+                      >
+                        {myitem.preferredCollectStartTime !== null && (
+                          <Typography
+                            variant='body2'
+                            color='textPrimary'
+                            gutterBottom
+                            align='left'
+                          >
+                            {myitem.preferredCollectStartTime}
+                          </Typography>
+                        )}
+                      </GridItem>
+                    </GridContainer>
+                  </GridItem>
+                </GridContainer>
               </GridItem>
-              <GridItem align='right'>
+              <GridItem xs={12} align='right'>
                 <Button>
                   <EmojiEvent />
                 </Button>
